requirements.txt
numpy
pandas
matplotlib
tensorflow
keras
opencv-python
Pillow
scikit-learn
streamlit
pyenv
Pipenv
Poetry